+++ Assets/ViveGrip/Scripts/ViveGrip_Interactable.cs
@@ -6,9 +6,7 @@ public class ViveGrip_Interactable : ViveGrip_Highlight {
 
   // These are called this on all scripts in the attached object:
 
-  // When touched and the interaction button is pressed
+  // When touched and the interaction button is pressed and released, respectively
   //   void ViveGripInteractionStart(ViveGrip_GripPoint gripPoint) {}
-
-  // Every frame when touched and the interaction button is held
   //   void ViveGripInteractionStop(ViveGrip_GripPoint gripPoint) {}
 }